- python.prof interface - Supports float as returns

+++ pymodules/ZBX_time.py
@@ -1,3 +1,3 @@
 import time
 def main(*args):
-	return long(time.time())
+	return time.time()
